I think this might actually be correct...

+++ testloop.sh
@@ -4,7 +4,10 @@
 #and the testing of those files with the checkfile.py script. I let this run for hours to
 #check that the program is working correctly. If either ./checkfile returns a failure
 #or the POC gets stuck, we know it's not correct.
-
+gcc -o poc poc.c -lpthread
+if [ $? -ne 0 ]; then
+	exit
+fi
 quitting=0 
 rm file*
 while [ $quitting -eq 0 ]; do
